Move get_config method
